safer typing for stride: int -> size_t
first pass, may have missed a bunch.

even if picture dimension (width/height) is reasonable, 'stride' can be
arbitrary large. 'size_t' is the natural type here, and makes calculations
like 'ptr + y * stride' safe.

I left WebPCopyPlane out of this patch's scope for now

+++ src/dsp/alpha_processing_sse41.c
@@ -19,9 +19,9 @@
 
 //------------------------------------------------------------------------------
 
-static int ExtractAlpha(const uint8_t* argb, int argb_stride,
+static int ExtractAlpha(const uint8_t* argb, size_t argb_stride,
                         int width, int height,
-                        uint8_t* alpha, int alpha_stride) {
+                        uint8_t* alpha, size_t alpha_stride) {
   // alpha_and stores an 'and' operation of all the alpha[] values. The final
   // value is not 0xff if any of the alpha[] is not equal to 0xff.
   uint32_t alpha_and = 0xff;
